MB-54540 Put Limit after early Order only when index spans are exact
In previous implementation of Early Order, we always put the Limit
operator right after the early Order operator. This could be
problematic in cases where there are additional filters to be applied
and thus could discard results after the Order/Limit operation, which
will result in less number of results returned.

We already have checks for "exact" index spans before applying early
Offset operator. Extend this logic to also apply to the early Limit
operator, such that only when the index spans are exact we apply the
limit early, otherwise only the Order is early, and the Limit is
applied at the end, in which case we'll rely on early termination
once the Limit is satisfied, to stop the plan execution and potentially
save on number of documents fetched.

Change-Id: Ib2c41f9d4d5cace5d99b91b36993a06a420c0f32
Reviewed-on: https://review.couchbase.org/c/query/+/182742
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>