MB-49506 : GROUP AS - Introduce GROUP AS clause as an optional clause in GROUP BY
GROUP AS allows us to preserve the documents in every group. The GROUP AS field in the output of a group will be an array of all the individual documents in the FROM clause that belong to that group.

Reference in Analytics - https://docs.couchbase.com/server/current/analytics/3_query.html#Group_As_clauses

Change-Id: Ica9c9d73bede36c6b344afa3bb9f967d7b4cb8fb
Reviewed-on: https://review.couchbase.org/c/query/+/182861
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>