MB-54568 Fix logic in choosing primary index and leadkey-missing index
With RBO (rule-based optimizer) when comparing two indexes, if one
of them is a primary index, and other other one is a leadkey-missing
index, and there are no sargable index keys (basically the
leadkey-missing index is being used as a primary index), then pick
the primary index since it is smaller. However the current check is
flawed in that it results in favoring whatever index comes in first
if sargable index keys exist in either of the index.

Fix the code logic to achieve the original intent.

Change-Id: I9b5629adcaca1e7524873eb13adfc691ce4dd61d
Reviewed-on: https://review.couchbase.org/c/query/+/182878
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>