MB-54045 Do not use join predicates as index span if under hash join
When generating an index span for an index scan under hash join, we
should not use join predicates since we cannot evaluate expressions
from the other keyspace yet. This behavior is correct in 6.5 and 6.6,
however, in 7.0 and above, with the fix for MB-30813, in order to
allow correlation variables to be used in an index span, we added
an additional check for whether the query block is correlated. The
issue with that is the check does not properly distinguish between
a correlation variable and a reference to a different keyspace in
the same query block.

The fix is to add a check in case of correlated subuqery, to see
whether the expression references any other keyspaces in the same
query block.

MB-54045 Proper formalization for ExpressionTerm and SubqueryTerm under ANSI join

When formalizing an ExpressionTerm or a SubqueryTerm, currently a new
formalizer is always created and returned. A new formalizer is necessary
to check whether there is any correlation in the expression for ExressionTerm
or the subuqery for SubqueryTerm. However, if this ExpressionTerm or SubqueryTerm
is under an ANSI join/nest operation, we need to use (and return) the parent's
formalizer instead of the newly created formalizer. This logic already exists
for KeyspaceTerm, however it is not done for ExpressionTerm and SubqueryTerm.
In the problematic query two ExpressionTerms are used in a join operation, and
after formalizing both, the formalizer only contains the second term as
"immediate" term, and thus the join expression with the first term is incorrectly
detected as "correlated".

The fix is to check and use the parent's formalizer in case of ANSI join/nest,
similar to what's already done for KeyspaceTerm.

MB-54045 Do not consider primary index for index scan inside correlated subquery

We don't allow primary index scans for a keyspace inside a correlated
subquery (for performance reasons). Current code would add primary
index to the list of indexes to be considered, and it is possible
that when CBO is on we could prefer a primary index due to lower
cost. Later on we detect primary index and return an error, even
though a usable secondary index exists.

The fix is to exclude a primary index for a keyspace inside a
correlated subquery.

Change-Id: Ie5d9bccc9a8c118b325a3837d3de08660347cb45
Reviewed-on: https://review.couchbase.org/c/query/+/183112
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>