MB-54041 Do not allow hash join under the inner of nested-loop join
Hash join should not be under the inner of a nested-loop join
since the inner side of nested-loop join requires the "reopen"
logic that is not implemented for hash join.

The only chance this could happen is when a SubqueryTerm is
used on the inner side of a join, and the join is using
nested-loop join for whatever reason (e.g. in case of
correlation in which case hash join is not feasible). Then
inside the SubqueryTerm a join could be using hash join.

The fix is to mark the builder when the inner of a
nested-loop join is being planned, and avoid consideration
of hash join in such cases.

MB-54041 Use ExpressionScan for SubqueryTerm if under inner of nested-loop join

When a FROM clause subquery (SubqueryTerm) is used under the inner
side of a nested-loop join, use an ExpressionScan on top of the
subquery instead of embedding the subquery plan under the inner
of nested-loop join. This way, the "reopen" logic only needs to
work properly for ExpressionScan and we don't need to implement
"reopen" for all operators.

An added benefit is that if the subquery is not correlated, then
the result of the subquery will be cached and thus not executed
repeatedly under nested-loop join (which will be the case if the
subquery plan is embedded in the parent query plan).

If the subquery is correlated, then the ExpressionScan will
automatically re-execute the subquery.

A few additional things that need to be taken care of:

1. From clause subqueries do not appear in the "~subqueries"
   section of the explain since it is normally inlined. If we
   now put an ExpressionScan on top of the subquery, we need
   to add the subquery and its plan to "~subqueries" section.

2. If there are any optimizer hints in the SubqueryTerm,
   previously these hints are gathered into the parent
   query's optimizer hints and displayed together with
   the parent query. Now if we have a separate section
   under "~subqueries" for this subquery, its optimizer
   hints will be displayed in that section, and thus we
   also need logic to remove these hints from the parent
   query's optimizer hints.

3. The ExpressionScan itself can cache result of the
   expression evaluation if not correlated. In case we
   have a subquery as the expression, the subuqery
   evaluation also caches results, thus we don't need to
   cache the results again at ExpressionScan. Added logic
   to avoid caching result at ExpressionScan in this case.

4. With this fix, the previous restriction of not allowing
   hash join under inner of nested-loop join is obsolete,
   reverse part of previous fix.

MB-54041 Handle NL join under Hash join

Previous fix did not properly handle the case where a
nested-loop join could fall under the build side of a
hash join, and thus could miss a SubqueryTerm under
there.

Also traverse down the build side of a hash join/nest
when looking for ExpressionScan on top of SubqueryTerm.

Change-Id: Ie58bdc26b97fe5ccce01e426e03f62e0dedbdea5
Reviewed-on: https://review.couchbase.org/c/query/+/183144
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>