Use flags for order and group operator attributes
Specifically instead of spilling and clipping being indicated by
separate values, use flags to represent them.  This has the benefit of
better using the flags field for order by (introduced for early order)
thereby reducing the size of operators in plan text.  Group by is for
consistency with order by.

Change-Id: I5314f52aab45001e319819caec427fa76f63a5ed
Reviewed-on: https://review.couchbase.org/c/query/+/183182
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>