MB-54804 Do not include SELF as recommended index key
In Elixir SELF is no longer allowed as an index key (MB-51996).
To be consistent, we now also skip SELF as an index key in
recommended indexes from the index advisor.

When collecting index key information from predicates and others,
skip "self". Note only "self" is excluded, it does not exclude
other references to "self", such as "self[1]", "self[1:3]", etc.

Change-Id: I85bbed3f87c0f9e071b20fb6dda346d48fd8ff20
Reviewed-on: https://review.couchbase.org/c/query/+/183875
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>