MB-54809 Handle nested simple functions in index advisor
The index advisor attempts to advise on simple arithmatic
expressions, and if a query parameter or a volatile
expression is used, it attempts to separate the individual
components of the expression and advise on those.

The issue with this bug is that when we have a nested
simple functions, in this case, a substraction followed by
a division, current code does not handle properly when
an operand of the simple function may return an array
(indicating the operand is being handled as "separated").

The fix is to keep track of potential "separated" children,
and include these as "separated" current operands.

Change-Id: I488f5766247b8de557fdb256c49c705e5fbcd047
Reviewed-on: https://review.couchbase.org/c/query/+/183922
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>