MB-54859 Properly set exact span for index with leading missing key and a where clause
If an index has a WHERE clause (partial index), and also
leading missing key, a query with a WHERE clause that matches
the index WHERE clause and a LIMIT clause should be able
to push down the LIMIT, however currently LIMIT is not
pushed down.

The reason is current code kind of assumes an index with
leading missing key does not have a WHERE clause, and thus
the presence of the WHERE clause in the query prevents
the index scan from being set as "exact", even when the
query WHERE clause matches exactly the index WHERE clause.

The fix is for an index with leading index key, when a
WHOLE_SPAN is being generated (no other predicates), allow
the index scan to be set to "exact" if the query WHERE
clause matches the index WHERE clause.

Change-Id: Ie02f0ba6d357b1061cdcdd835235bab38d51bc09
Reviewed-on: https://review.couchbase.org/c/query/+/184057
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>