MB-54871 Use dummy base for Advise
The fix for MB-54740 uses a nil base pointer in NewOpContext(),
assuming there is no need for base during advise. In cost-based
advise, we run UPDATE STATISTICS as part of the Advise, and if
the keyspace is existing and non-empty, it uses random scan
for sampling and it actually uses the base in operator context,
which results in panic.

The fix is to use a dummy base (newBase()) instead of a nil
pointer in NewOpContext().

Change-Id: I04199906f8964066ef702fe61c91c6e0ec888d0e
Reviewed-on: https://review.couchbase.org/c/query/+/184110
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>