MB-54880 Allow recommended index with leading missing key to have a WHERE clause
The index advisor uses INFER to deduce whether a partial index should
be recommended. For a recommended index without leading missing key,
the WHERE clause is added to the recommended index. If there are no
index keys for the recommended index, it'll "convert" the WHERE clause
condition to an index key, e.g., for a WHERE clause condition of
  "type = 'hotel'"
it'll recommend an index on (type).

The logic for recommending an index with leading missing key follows
similar logic, except that the "conversion" happens before we check
whether a leading missing key can be recommended, as a result, we never
recommend an index with leading missing key that also has a WHERE clause.

We now switch the position of code such that we check for potential
leading missing key first, and if that's available, generate an index
recommendation with leading missing key that could have a WHERE clause.
When no such index keys exist, then we perform the "conversion".

Change-Id: I54a8605b8ba3f6c679ffeb9df038a24b1bcd590a
Reviewed-on: https://review.couchbase.org/c/query/+/184152
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>