MB-54881 Proper check for SYSTEM namespace in dictionary
For SYSTEM namespace, the Id() and Name() calls used to return
the same '#system'. With MB-51988 the Id() call remains the same
but the Name() call now returns 'system' (no leading '#' anymore).

The function algebra.IsSystem() checks for leading '#system',
which now no longer works if the name is missing leading '#',
which is the case for dictionary code since this is what
QualifiedName() returns for system keyspaces. This results in
system keyspaces not properly excluded.

The fix is to have an extra algebra.IsSystemName() call which
checks for the name without the leading '#', and use the new
call in dictionary code.

Change-Id: I1bffe7e451b5780555d8358868a67ab7f86ccd1b
Reviewed-on: https://review.couchbase.org/c/query/+/184171
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>