MB-55150 Request logging/tracing.
Add per-request tracing that is returned with the request results.  This
is to better support diagnostics in a serverless environment.

It is deliberate intention that the timestamps omit timezone and the
message level is a single character - these both help reduce the size
for better on-screen legibility (without meaningful data loss).

It is inevitable that if both request and server tracing are on the
anonymous function is invoked twice or the string is formatted twice.
A feedback value could be implemented to avoid this, but I figured the
overhead shouldn't be unexpected.

The 64 MiB per-request trace limit should easily be handled since it is
in temp files which can spill.  With 192 active requests it would be 12
GiB at most which shouldn't pose a problem on a server that can handle
192 servicers.

Use of the logging package functions with the additional argument
permits choice of logging location.  If logging is desired that
can't/won't be written to the query.log ever then Context logging
functions can be used directly.

The extension of the logging.Logger interface permits inclusion of
future logger implementations that will interface with the request level
logging readily.

Change-Id: I3d402100f9039c6bd99ac272a842546115e0c833
Reviewed-on: https://review.couchbase.org/c/query/+/184824
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>