MB-54952 Allow WHOLE_SPANS to be generated for leading array index key
When an array index key is in simplified form, e.g. ALL arr1, then when
we determine sargability we simply compare the binding expression of the
ANY clause with the ALL expr, e.g. ANY v IN arr1 SATISFIES ... END, when
arr1 expression matches we determine it is sargable, without checking the
SATISFIES clause. This could be problematic if the SATISFIES clause is
not actually sargable.

In this bug, we have a IS NOT VALUED clause in the SATISFIES clause,
which makes it not sargable. We return an error saying a sargable
predicate is not sarged.

In previous versions (e.g. 6.6.x, 7.0.4, 7.1.0) we generate a
_WHOLE_SPANS in this case, i.e. the entire index is scaned. For
backwards compatibility we now also generate a _WHOLE_SPANS in this
special case if the array index key is the leading index key.
(if the array index key is not leading, then it sort of does not matter).
This is a "temporary" fix, meant to be ported to earlier versions.
A more relevant fix will be done for Elixir/Morpheus later.

Change-Id: I30562e7a6eeeda9e2470a6c487e2d7960dba2de0
Reviewed-on: https://review.couchbase.org/c/query/+/184482
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>