MB-54952 Proper determination of sargability for array index in simplified form
For an array index key specified in simplified form, e.g. ALL arr1,
currently when determining sargability we simply check whether the
ANY clause matches the array specification (arr1); in contrast, for
an equivalent array index key specification:
 ALL ARRAY v FOR v IN arr1 END
we check for both the array specification (arr1) as well as
sargability of the SATISFIES clause. This results in inconsistency
for something like:
  ANY v IN arr1 SATISFIES v IS NOT VALUED END
which is sargable if the array index key is in simple form, but
not sargable if the array index key is in the full ARRAY syntax.

The fix is to also check the SATISFIES clause for the simple form
as well. Note this may result in a predicate no longer able to
use an index scan (previously a _WHOLE_SPANS is generated). We
fall back to sequential scan in such cases. (This fix will not be
backported to earlier versions).

Change-Id: I21ee3f6085cfd02f6e6f608d0a54e4f18db98822
Reviewed-on: https://review.couchbase.org/c/query/+/184547
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>