MB-48675 Refine fix
Lowers the minimum result-cap size to 512 KiB.

Silently limits result-cap to available quota when quota is in effect -
this is so we see a "Response size " error ahead of a quota exceeded
error.  It also means that requests with small enough responses don't
fail up front because an over-large result-cap was specified.

Change-Id: I612331c9da15ac679d81f4ba72611c7a06c7b647
Reviewed-on: https://review.couchbase.org/c/query/+/184671
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>