MB-44279 Formalization changes for CTE
Previously if a CTE is used, it is first pushed as a binding
variable in the formalizer such that it is "allowed" as a
reference anywhere in the SELECT query.

Normally for any FROM clause term we add it to the "allowed"
list when it is first used, then it is available for any
subsequent clauses, and if the same alias is used subsequently
we can detect that and raise an error. This mechanism does not
work for a CTE variable since it becomes "allowed" when the
CTE definition is encountered, before any actual references to
the CTE variable. To make subsequent references work we had to
disable the "duplication" checking and thus the same mechanism
for detection of duplicated aliases does not work for CTE
variables.

To fix this issue, we need to change the way CTE variables are
handled in the formalizer, i.e. we also need to only add a CTE
variable to be "allowed" when it first appears in a FROM
clause, this way similar duplicate detection will work for CTE
variable as well.

To make this work, when CTE definition is encountered, we only
put the CTE information in the "withs" field in the formalizer,
which we already do previously, but now we need to keep track
of a few additional information, as a result a new WithInfo
structure is defined and populated as a CTE is being processed.

We also need to properly "inherit" CTE information (withs) when
a new formalizer is being constructed based on an existing
formalizer (the parent). When a full select is being handled,
we need to mark whether it is a subquery, and if it is, need to
inherit the parent's CTE info.

If a CTE is correlated, instead of relying on the formalizer
itself to keep this info (since the CTE is pushed to the
formalizer early on), we now keep this info in the new WithInfo
structure itself, and transfer the correlation info as needed
when the CTE is referenced.

We also support referencing a CTE as a general expression, i.e.,
the CTE does not need to be referenced in the FROM clause first
before it can be references as an expression. To make this work
when an identifier is formalized we need to check and handle if
the identifier is a CTE reference.

Change-Id: Iabb53f90d6c03d7cab8efd84a9bbb57e22bad5f1
Reviewed-on: https://review.couchbase.org/c/query/+/185172
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>