MB-55236 Handle index selection better for OR clauses in RBO
In the problematic scenario we have two indexes defined:
  ix1 (c1, c2, cdate) where type = "doc"
  ix2 (c1, c2, c3, ldate desc) where type = "doc"

The query predicate:
  type = "doc" AND c1 = 10 AND (c2 = "xyz" OR c3 = "abc")

Both ix1 and ix2 are sargable in this case, and when a predicate
contains OR clause, we use/set the maximum of the OR subarms for
nSargKeys which could potentially be problematic if different
subarms has different nSargKeys, and the current code uses
nSargKeys for comparisons to determine whether one index is
"better" than another. To make this work better we pass in an
additional indicator if an OR clause is being used, in which
case we use minKeys instead of nSargKeys for certain comparisons.
Another potential issue is for this query we could use ix2 for
early order, thus add comparison for checking index flags at the
same places where we currently check for index pushdown properties.

A different issue is that when we use individual arms of the OR
subclause to construct subarms of a UnionScan, we currently take
the minimum of the sargLength from all subarms, this is somewhat
inconsistent with the case where the entire OR clause is used since
in that case we take the maximum of the sargLength from all subterms
of the OR clause. To make this a more fair comparison, we also take
max instead of min of the sargLengths from individual subterms of
the OR clause when we are comparing a plan using the entire OR
clause vs a plan with individual subterms of the OR clause.

Change-Id: I7fad043f4bdde94e88042d5acf09ca5e70349902
Reviewed-on: https://review.couchbase.org/c/query/+/185505
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>