MB-55236 Revise fix, proper accounting for OR clause in sargability
There is an additional issue in how we account for sargable
index keys for OR clauses.

In the problematic scenario we have two indexes defined:
  ix1 (c1, c2, cdate) where type = "doc"
  ix2 (c1, c2, c3, ldate desc) where type = "doc"

The query predicate is:
  type = "doc" AND c1 = 10 AND (c2 = "xyz" OR c3 = "abc")

It turns out that the number of sargable keys we have depends
on the order of the subterms in the OR clause. If c2 is first
as above then we have nSargKeys = 1, if we switch the two:
  type = "doc" AND c1 = 10 AND (c3 = "abc" OR c2 = "xyz")
then nSargKeys = 2. This is obviously incorrect.

When we determine sargability of an OR clause, instead of taking
the skeys[] array from an arm of the OR clause as we do now
(which is what introduced the inconsistency), we now keep a
running count of the sargable keys, and an index key is only
considered sargable if it is sargable for all subterms of the
OR clause, which is consistent with the actual spans generated
since if it's not included in one of the subterms, a _WHOLE_SPANS
would have been generated for that index key.

With this change, the sargLength is effective now the min of
all subterms of OR, thus we can change the orSargLength back
to min (instead of max). We also no longer need the special
case of isOr.

Change-Id: Ife86477716f8ce6ce1f8b0d6b7a967efa97d2f9c
Reviewed-on: https://review.couchbase.org/c/query/+/185564
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/185570
Well-Formed: Restriction Checker