MB-55321 Improve distinct scan cost
A DistinctScan is needed in the query plan if an array index
is in use (due to potential duplicated document keys).
Current costing for DistinctScan does not distinguish whether
the actual array index key is used in the index span or not.
In case where an array index is being used but the array
index key is not part of the index span, costing should be
simpler and no need to actually look at the distributions of
the array index key.

Change-Id: I8c573665efaf403a457f77a7cfc1b12dcf4d7173
Reviewed-on: https://review.couchbase.org/c/query/+/185870
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>