MB-51114 Limit number of spans generated from a long IN-list
The planner has an upper limit of number of index spans generated
to avoid potential performance issues in the index scan. However
currently this limit is not enforced if an IN-list has too many
elements, resulting in number of spans exceeding the limit.

Add logic in handling of IN-list to check the number of elements
in the IN-list, and if it exceeds the limit, generate a single
index span with ARRAY_MIN/ARRAY_MAX as the bounds, and apply the
IN-list predicate post index scan.

Change-Id: Id47d326342ab1a85beea2a5c64e23012e4a4b0ce
Reviewed-on: https://review.couchbase.org/c/query/+/186516
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>