MB-55547 Check for order of primary scan when building ORDER plan during join enumeration
When CBO is in effect, during join enumeration, if the query has an
ORDER BY clause, we attempt to build ORDER plan with each keyspace
such that we could choose a plan with proper order provided to avoid
a SORT at the end. During this process, proper order is checked for
all secondary indexes. However, if there is no appropriate secondary
indexes, we also attempt to build a primary scan, with a
primary index. The issue is that such a primary
scan is incorrectly marked as providing order and later on we don't
build a SORT operator, which results in wrong result.

The fix is to properly check for ordering of such primary scans.

Change-Id: I5b8066ed0e9a8b6326d322a9b849cc3c2f4d63ec
Reviewed-on: https://review.couchbase.org/c/query/+/186582
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/186583
Well-Formed: Restriction Checker