MB-55559 Do not rearrange first scan under OrderedIntersectScan
Part of the fix for MB-51288 sorts the index scans under
IntersectScan and OrderedIntersectScan by cost if using
CBO, for explain stability. The issue here is that for
OrderedIntersectScan the first scan is what provides the
ordering and thus should not be changed.
Avoid changing the first scan under OrderedIntersectScan.

Change-Id: I2845688ee31f3230b67577ea4cd7bde7de8ef068
Reviewed-on: https://review.couchbase.org/c/query/+/186658
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>