MB-55618 Limit blocking when writing responses
Requests are added to monitoring when created and only removed when
destroyed (not pooled).  Monitoring is effected with a remaining duration
counter in each request - reset for each I/O write operation - that is
decreased each loop of a single monitor routine.  The monitor routine
loops on a 10 second ticker making this the maximum precision of the
timeout mechanism.  Furthermore, to limit overhead remaining duration
counter operations are not synchronised; the worst side effect is
essentially increasing the minimum timeout by 10 seconds.

Change-Id: Ia649be2de8c362a2c6870cc8d70b2ee96426d3c8
Reviewed-on: https://review.couchbase.org/c/query/+/186882
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>