MB-55560 Properly handle non-pushable ON-clause during join enumeration
For ANSI JOINs we gather ON-clauses from the inner joins and
"push" the ON-clause predicates to individual base keyspaces
since the ON-clause from inner joins behave like WHERE clause
predicates. However, if there is an outer join in the query,
then even the ON-clause from a subsequent inner join may not
be "pushable" if it references the outer keyspace (and is not
null-rejecting, since if it is we would have transformed the
outer join to an inner join). In such situations the ON-clause
is not pushed to the individual keyspaces and is only
evaluated when the join is being considered.

With join enumeration, we could potentially change the order
of evaluation of the joins, and current code does not
properly handle such non-pushable ON-clause predicates
which could lead to wrong result.

Add logic to mark predicates involving outer keyspaces in
such non-pushable ON-clauses as non-pushable, such that:
  - it is not used for index selection for a keyspace
  - it is gathered as a post-join filter (since it is
    originally evaluated after the outer join, and thus
    can be treated as post-join), this makes sure that
    the predicate is evaluated eventually

Change-Id: I52130a58b09bbac515e95a2ad3b69276dfbd7896
Reviewed-on: https://review.couchbase.org/c/query/+/186905
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/186910
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker