MB-55658 add a goroutine pool to avoid growing the stack of each operator
I have found another 6K in Q1 by adding a pool of goroutines for operators and scans.
The logic is that by having goroutines live longer, they grow their stack and lower the contention on
moreStack().
You still have to grow the stack when the goroutine is initially fired, and also you have to avoid anonymous functions
because they get allocated on the heap no matter what, which eats up half of the advantage.
Sadly, there must be another major bottleneck in Q2, because I only get 2.5K out of this change.

Change-Id: I3015b80eb4a19cd41a13860272284667371326fd
Reviewed-on: https://review.couchbase.org/c/query/+/186965
Tested-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>