MB-55663 Better marking of join hint specified on the first FROM term
With legacy (embedded) join hints if a join hint is specified on
the first FROM term, an error will be returned. With the new
relational-style hints, if a join hint is specified on the first
FROM term, with join enumeration the join hint will be
considered when the first FROM term is considered for a join.
If join enumeration is not done, currently we only report the
join hint as "NOT FOLLOWED".

Mark the join hint with a error that is similar to existing
error in such cases. It is still however only a marked error
on the directive itself rather than an error returned for the
query (which is the case for legacy join hint).

Change-Id: Ife266a28fe41d1016fa5cd4b6dc3d9dc841e7bec
Reviewed-on: https://review.couchbase.org/c/query/+/187149
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>