MB-55463 More processing of GSI errors.
GSI returns some errors for transient conditions which are automatically
handled in the background by GSI.  This change allows reporting of them as
warnings with HTTP status 202 (accepted) rather than as errors with
status 500.

Further to this, GSI errors that can be are broken down into separate
fields that populate the cause (reported as "reason").

Change-Id: Icecbc5c939e7739d2c24198e2047ff0ba7db5a95
Reviewed-on: https://review.couchbase.org/c/query/+/188218
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>