MB-55757 Fix cost comparison of Union Scan vs non-Union Scan
When CBO is in effect, for OR predicate we first try to generate
a regular index scan, and also a Union Scan, and compare the
costs to pick one. This cost comparison should include the cost
for Fetch as well since it's possible/likely that the two scans
will produce different cardinality.

Change-Id: I3e50d59ee0d70a0691b13bebd09d6d849efb379a
Reviewed-on: https://review.couchbase.org/c/query/+/187600
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>