MB-55849 Limit filter evaluation for UNNEST at runtime
For inner UNNEST, the planner adds two additional filters:
  UNNEST <expr> AS <alias>
  - IS_ARRAY(<expr>)
  - <alias> IS NOT MISSING
the first (IS_ARRAY) is meant to eliminate at the original
document level any document that does not have an array as
unnest expression; the second is for index selection
purpose.

The issue here is that if the unnest expression is an
expensive function (e.g. a UDF) then the expression is
evaluated twice, once as a filter and once during unnest.

To fix this, we add a check and if we know <expr> is
producing an array (e.g. certain built-in functions)
then there is no need to add the IS_ARRAY filter. For
other situations (e.g. UNNEST p.lineitems) we will
preserve the derived IS_ARRAY filter.

For the IS NOT MISSING filter, there is no need to
reevaluate the filter after the unnest. Remove this
filter from unnest filter.

Change-Id: Ic3c81a5ffa39449502de74f52429c65ec0e33312
Reviewed-on: https://review.couchbase.org/c/query/+/187801
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>