MB-55849 Revise fix
Turns out the IS_ARRAY filter is also used for index
selection, if the index contains an entire array as
the index key. Thus we cannot remove the IS_ARRAY
filter.

Modify the fix such that the IS_ARRAY predicate is
added, but it is handled similarly to the
IS NOT MISSING filter for the unnest alias: we skip
the filter when gathering filters such that it does
not get evaluated as a separate filter.

Change-Id: I97980c6edbaf07e0f5460e4f40e286c0d78fcd82
Reviewed-on: https://review.couchbase.org/c/query/+/187807
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>