MB-56041 Add handling for additional format variants.
Observed the date format with a space before TZ information in
travel-sample so this adds handling to deal with such non-standard
jformats.

(As this adds no fixed standard formats, existing processing that
relies on matching one of the known formats is unaffected.)

This also simplifies the interrogation used to determine format.

Change-Id: I8b2960f2b9ef18ab3a1a70f8d0f6eadaac396871
Reviewed-on: https://review.couchbase.org/c/query/+/188039
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>