MB-55869 Ignore correlated subquery index error inside SubqueryTerm during join enumeration
When a SubqueryTerm is being planned during join enumeration,
it's possible to get an error of missing secondary index for
correlated subquery. We can ignore this error during scan
plan generation since we will consider joins later, and it
is possible to get a plan with nested-loop join. This is
similar to the way we currently ignore primary index not
found or not online errors during join enumeration.

Change-Id: I458186a9f906d5184a5f32852f007d18f4ecbe4e
Reviewed-on: https://review.couchbase.org/c/query/+/188058
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>