MB-55869 Do not duplicate derived optimizer hints inside SubqueryTerm during join enumeration
When the subquery inside a SubqueryTerm is being planned, it first
derive new optimizer hints from the legacy style (embedded) hints
and add the new hints to the algebra node for the subquery. With
join enumeration, when the subquery is planned multiple times, do
not try to derive the same set of new optimizer hints multiple times.

Change-Id: If180828f603fe177eec2d1067ca99cb2681d5eeb
Reviewed-on: https://review.couchbase.org/c/query/+/188101
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>