MB-55978 Do not inherit Covers from parent in ExpressionScan
When an AnnotatedValue is being created from a ScopeValue, if
the parent of the ScopeValue is an AnnotatedValue, we
automatically inherit the Covers from the parent AnnotatedValue
to the new AnnotatedValue, in effect we carry over the Covers.

The issue with this defect is when we are doing a nested-loop
join, when processing the inner of the join, the parent includes
document value from the outer of the join. If the inner is an
ExpressionScan (e.g. a CTE) then we currently generate a value
from the ExpressionScan that also includes the Covers
information from the parent. In case the CTE is not correlated,
we cache the value from the ExpressionScan, and on subsequent
joins we end up using the cached value which includes Covers
from the very first outer document, giving wrong result.

The fix is to mark and detect when the ExpressionScan is under
a nested-loop join, and remove Covers information from the
generated AnnotatedValue. This is consistent with the case
where an index scan is under the inner of a nested-loop join
where we do not use the parent value. SubqueryTerm does not
need to be handled differently since in case the SubqueryTerm
is under the inner of a nested-loop join we build an
ExpressionScan on top of the subquery expression.

Change-Id: I45edc9b504551745a157832c0aceea6cb1b6573a
Reviewed-on: https://review.couchbase.org/c/query/+/188230
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/188241
Well-Formed: Restriction Checker